# Common.Contracts/openapi_v3.yaml
openapi: 3.0.4
info:
  title: LPshows.live API v3
  description: This is the API for the Linkin Park Concert Calendar fan project
  version: 3.0.0
servers:
  - url: /
  - url: http://localhost:5133
paths:
...
  /v3/countries:
    post:
      tags:
        - Countries
      summary: Creates a new country
      operationId: CreateCountry
      requestBody:
        description: Data of the new country
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCountryRequest"
          text/json:
            schema:
              $ref: "#/components/schemas/CreateCountryRequest"
          application/*+json:
            schema:
              $ref: "#/components/schemas/CreateCountryRequest"
        required: true
      responses:
        "200":
          description: ""
        "401":
          description: "Unauthorized: Credentials are missing or not valid"
        "403":
          description: "Forbidden: Credentials are valid, but the caller is not allowed to
            perform this operation"
      x-amazon-apigateway-integration:
        type: HTTP_PROXY
        httpMethod: POST
        uri: https://api.dev.lpshows.felixsfd.de/v3/countries
      security:
        - Bearer: []
    get:
      tags:
        - Countries
      summary: Returns a list of all countries
      operationId: GetCountries
      responses:
        "200":
          description: List of all countries
          content:
            text/plain:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Country"
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Country"
            text/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Country"
      x-amazon-apigateway-integration:
        type: HTTP_PROXY
        httpMethod: GET
        uri: https://api.dev.lpshows.felixsfd.de/v3/countries
  /v3/countries/{countryCode}:
    put:
      tags:
        - Countries
      summary: Updates an existing country
      operationId: UpdateCountry
      parameters:
        - name: countryCode
          in: path
          description: ISO-code of the country
          required: true
          schema:
            type: string
      requestBody:
        description: new data of the country
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCountryRequest"
          text/json:
            schema:
              $ref: "#/components/schemas/UpdateCountryRequest"
          application/*+json:
            schema:
              $ref: "#/components/schemas/UpdateCountryRequest"
        required: true
      responses:
        "200":
          description: the updated data
          content:
            text/plain:
              schema:
                $ref: "#/components/schemas/Country"
            application/json:
              schema:
                $ref: "#/components/schemas/Country"
            text/json:
              schema:
                $ref: "#/components/schemas/Country"
        "401":
          description: "Unauthorized: Credentials are missing or not valid"
        "403":
          description: "Forbidden: Credentials are valid, but the caller is not allowed to
            perform this operation"
      x-amazon-apigateway-integration:
        type: HTTP_PROXY
        httpMethod: PUT
        uri: https://api.dev.lpshows.felixsfd.de/v3/countries/{countryCode}
        requestParameters:
          integration.request.path.concertId: method.request.path.countryCode
      security:
        - Bearer: []
    get:
      tags:
        - Countries
      summary: Returns a country by its ISO code
      operationId: GetCountryByIsoCode
      parameters:
        - name: countryCode
          in: path
          description: 3-letter ISO-code of the country
          required: true
          schema:
            type: string
      responses:
        "200":
          description: ""
          content:
            text/plain:
              schema:
                $ref: "#/components/schemas/Country"
            application/json:
              schema:
                $ref: "#/components/schemas/Country"
            text/json:
              schema:
                $ref: "#/components/schemas/Country"
      x-amazon-apigateway-integration:
        type: HTTP_PROXY
        httpMethod: GET
        uri: https://api.dev.lpshows.felixsfd.de/v3/countries/{countryCode}
        requestParameters:
          integration.request.path.concertId: method.request.path.countryCode
    delete:
      tags:
        - Countries
      summary: Deletes a country
      operationId: DeleteCountryByIsoCode
      parameters:
        - name: countryCode
          in: path
          description: 3-letter ISO-code of the country
          required: true
          schema:
            type: string
      responses:
        "200":
          description: no content
        "401":
          description: "Unauthorized: Credentials are missing or not valid"
        "403":
          description: "Forbidden: Credentials are valid, but the caller is not allowed to
            perform this operation"
      x-amazon-apigateway-integration:
        type: HTTP_PROXY
        httpMethod: DELETE
        uri: https://api.dev.lpshows.felixsfd.de/v3/countries/{countryCode}
        requestParameters:
          integration.request.path.concertId: method.request.path.countryCode
      security:
        - Bearer: []
  /v3/countries/{countryCode}/states:
    post:
      tags:
        - Countries
      summary: Creates a new state in a country
      operationId: CreateState
      parameters:
        - name: countryCode
          in: path
          description: ISO code of the country where this state is located in
          required: true
          schema:
            type: string
      requestBody:
        description: Data of the new country
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateStateRequest"
          text/json:
            schema:
              $ref: "#/components/schemas/CreateStateRequest"
          application/*+json:
            schema:
              $ref: "#/components/schemas/CreateStateRequest"
        required: true
      responses:
        "200":
          description: ""
        "401":
          description: "Unauthorized: Credentials are missing or not valid"
        "403":
          description: "Forbidden: Credentials are valid, but the caller is not allowed to
            perform this operation"
      x-amazon-apigateway-integration:
        type: HTTP_PROXY
        httpMethod: POST
        uri: https://api.dev.lpshows.felixsfd.de/v3/countries/{countryCode}/states
        requestParameters:
          integration.request.path.concertId: method.request.path.countryCode
      security:
        - Bearer: []
    get:
      tags:
        - Countries
      summary: Returns a list of all states in a country
      operationId: GetStatesInCountry
      parameters:
        - name: countryCode
          in: path
          description: 3-letter ISO-code of the country
          required: true
          schema:
            type: string
      responses:
        "200":
          description: ""
          content:
            text/plain:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StateWithCountry"
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StateWithCountry"
            text/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StateWithCountry"
      x-amazon-apigateway-integration:
        type: HTTP_PROXY
        httpMethod: GET
        uri: https://api.dev.lpshows.felixsfd.de/v3/countries/{countryCode}/states
        requestParameters:
          integration.request.path.concertId: method.request.path.countryCode
  /v3/countries/{countryCode}/states/{stateCode}:
    put:
      tags:
        - Countries
      summary: Updates a state in a country
      operationId: UpdateState
      parameters:
        - name: countryCode
          in: path
          description: ISO code of the country where this state is located in
          required: true
          schema:
            type: string
        - name: stateCode
          in: path
          required: true
          schema:
            type: string
      requestBody:
        description: Data of the updated state
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateStateRequest"
          text/json:
            schema:
              $ref: "#/components/schemas/UpdateStateRequest"
          application/*+json:
            schema:
              $ref: "#/components/schemas/UpdateStateRequest"
        required: true
      responses:
        "200":
          description: ""
          content:
            text/plain:
              schema:
                $ref: "#/components/schemas/StateWithCountry"
            application/json:
              schema:
                $ref: "#/components/schemas/StateWithCountry"
            text/json:
              schema:
                $ref: "#/components/schemas/StateWithCountry"
        "401":
          description: "Unauthorized: Credentials are missing or not valid"
        "403":
          description: "Forbidden: Credentials are valid, but the caller is not allowed to
            perform this operation"
      x-amazon-apigateway-integration:
        type: HTTP_PROXY
        httpMethod: PUT
        uri: https://api.dev.lpshows.felixsfd.de/v3/countries/{countryCode}/states/{stateCode}
        requestParameters:
          integration.request.path.concertId: method.request.path.countryCode
          integration.request.path.stateCode: method.request.path.stateCode
      security:
        - Bearer: []
    get:
      tags:
        - Countries
      summary: Returns a state in a country
      operationId: GetState
      parameters:
        - name: countryCode
          in: path
          description: 3-letter ISO-code of the country
          required: true
          schema:
            type: string
        - name: stateCode
          in: path
          description: Code of the state
          required: true
          schema:
            type: string
      responses:
        "200":
          description: The state including the information about the country
          content:
            text/plain:
              schema:
                $ref: "#/components/schemas/StateWithCountry"
            application/json:
              schema:
                $ref: "#/components/schemas/StateWithCountry"
            text/json:
              schema:
                $ref: "#/components/schemas/StateWithCountry"
      x-amazon-apigateway-integration:
        type: HTTP_PROXY
        httpMethod: GET
        uri: https://api.dev.lpshows.felixsfd.de/v3/countries/{countryCode}/states/{stateCode}
        requestParameters:
          integration.request.path.concertId: method.request.path.countryCode
          integration.request.path.stateCode: method.request.path.stateCode
    delete:
      tags:
        - Countries
      summary: Deletes a state
      operationId: DeleteState
      parameters:
        - name: countryCode
          in: path
          description: 3-letter ISO-code of the country
          required: true
          schema:
            type: string
        - name: stateCode
          in: path
          description: Code of the state
          required: true
          schema:
            type: string
      responses:
        "200":
          description: no content
        "401":
          description: "Unauthorized: Credentials are missing or not valid"
        "403":
          description: "Forbidden: Credentials are valid, but the caller is not allowed to
            perform this operation"
      x-amazon-apigateway-integration:
        type: HTTP_PROXY
        httpMethod: DELETE
        uri: https://api.dev.lpshows.felixsfd.de/v3/countries/{countryCode}/states/{stateCode}
        requestParameters:
          integration.request.path.concertId: method.request.path.countryCode
          integration.request.path.stateCode: method.request.path.stateCode
      security:
        - Bearer: []
  /v3/countries/{countryCode}/cities:
    post:
      tags:
        - Countries
      summary: Creates a new city in a country
      operationId: CreateCity
      parameters:
        - name: countryCode
          in: path
          description: ISO code of the country where this city is located in
          required: true
          schema:
            type: string
      requestBody:
        description: Data of the new city
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCityRequest"
          text/json:
            schema:
              $ref: "#/components/schemas/CreateCityRequest"
          application/*+json:
            schema:
              $ref: "#/components/schemas/CreateCityRequest"
        required: true
      responses:
        "200":
          description: ""
        "401":
          description: "Unauthorized: Credentials are missing or not valid"
        "403":
          description: "Forbidden: Credentials are valid, but the caller is not allowed to
            perform this operation"
      x-amazon-apigateway-integration:
        type: HTTP_PROXY
        httpMethod: POST
        uri: https://api.dev.lpshows.felixsfd.de/v3/countries/{countryCode}/cities
        requestParameters:
          integration.request.path.concertId: method.request.path.countryCode
      security:
        - Bearer: []
    get:
      tags:
        - Countries
      summary: Returns a list of all cities in a country
      operationId: GetCitiesInCountry
      parameters:
        - name: countryCode
          in: path
          description: 3-letter ISO-code of the country
          required: true
          schema:
            type: string
      responses:
        "200":
          description: ""
          content:
            text/plain:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CityWithCountry"
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CityWithCountry"
            text/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CityWithCountry"
      x-amazon-apigateway-integration:
        type: HTTP_PROXY
        httpMethod: GET
        uri: https://api.dev.lpshows.felixsfd.de/v3/countries/{countryCode}/cities
        requestParameters:
          integration.request.path.concertId: method.request.path.countryCode
  /v3/countries/{countryCode}/cities/{cityId}:
    put:
      tags:
        - Countries
      summary: Updates a city in a country
      operationId: UpdateCity
      parameters:
        - name: countryCode
          in: path
          description: ISO code of the country where this state is located in
          required: true
          schema:
            type: string
        - name: cityId
          in: path
          required: true
          schema:
            pattern: ^-?(?:0|[1-9]\d*)$
            type: integer
            format: int32
      requestBody:
        description: Data of the updated city
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCityRequest"
          text/json:
            schema:
              $ref: "#/components/schemas/UpdateCityRequest"
          application/*+json:
            schema:
              $ref: "#/components/schemas/UpdateCityRequest"
        required: true
      responses:
        "200":
          description: ""
          content:
            text/plain:
              schema:
                $ref: "#/components/schemas/StateWithCountry"
            application/json:
              schema:
                $ref: "#/components/schemas/StateWithCountry"
            text/json:
              schema:
                $ref: "#/components/schemas/StateWithCountry"
        "401":
          description: "Unauthorized: Credentials are missing or not valid"
        "403":
          description: "Forbidden: Credentials are valid, but the caller is not allowed to
            perform this operation"
      x-amazon-apigateway-integration:
        type: HTTP_PROXY
        httpMethod: PUT
        uri: https://api.dev.lpshows.felixsfd.de/v3/countries/{countryCode}/cities/{cityId}
        requestParameters:
          integration.request.path.concertId: method.request.path.countryCode
          integration.request.path.cityId: method.request.path.cityId
      security:
        - Bearer: []
    get:
      tags:
        - Countries
      summary: Returns a state in a country
      operationId: GetCity
      parameters:
        - name: countryCode
          in: path
          description: 3-letter ISO-code of the country
          required: true
          schema:
            type: string
        - name: cityId
          in: path
          description: ID of the city
          required: true
          schema:
            pattern: ^-?(?:0|[1-9]\d*)$
            type: string
            format: uint32
      responses:
        "200":
          description: The city including the information about the country
          content:
            text/plain:
              schema:
                $ref: "#/components/schemas/CityWithCountry"
            application/json:
              schema:
                $ref: "#/components/schemas/CityWithCountry"
            text/json:
              schema:
                $ref: "#/components/schemas/CityWithCountry"
      x-amazon-apigateway-integration:
        type: HTTP_PROXY
        httpMethod: GET
        uri: https://api.dev.lpshows.felixsfd.de/v3/countries/{countryCode}/cities/{cityId}
        requestParameters:
          integration.request.path.concertId: method.request.path.countryCode
          integration.request.path.cityId: method.request.path.cityId
    delete:
      tags:
        - Countries
      summary: Deletes a city
      operationId: DeleteCity
      parameters:
        - name: countryCode
          in: path
          description: 3-letter ISO-code of the country where the city is located in
          required: true
          schema:
            type: string
        - name: cityId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: no content
        "401":
          description: "Unauthorized: Credentials are missing or not valid"
        "403":
          description: "Forbidden: Credentials are valid, but the caller is not allowed to
            perform this operation"
      x-amazon-apigateway-integration:
        type: HTTP_PROXY
        httpMethod: DELETE
        uri: https://api.dev.lpshows.felixsfd.de/v3/countries/{countryCode}/cities/{cityId}
        requestParameters:
          integration.request.path.concertId: method.request.path.countryCode
          integration.request.path.cityId: method.request.path.cityId
      security:
        - Bearer: []
...
components:
  schemas:
...
    CityWithCountry:
      required:
        - country
        - countryCode
        - name
        - nativeName
      type: object
      properties:
        country:
          $ref: "#/components/schemas/Country"
        state:
          oneOf:
            - type: "null"
            - $ref: "#/components/schemas/State"
        countryCode:
          type: string
        id:
          pattern: ^-?(?:0|[1-9]\d*)$
          type: string
          format: uint32
        stateCode:
          type: string
        name:
          type: string
        nativeName:
          type: string
...
    Country:
      required:
        - isoCode
        - name
        - nativeName
      type: object
      properties:
        isoCode:
          maxLength: 3
          minLength: 3
          type: string
        name:
          type: string
        nativeName:
          type: string
    CreateCityRequest:
      required:
        - name
        - nativeName
      type: object
      properties:
        stateCode:
          type: string
        name:
          type: string
        nativeName:
          type: string
...
    CreateCountryRequest:
      required:
        - isoCode
        - name
        - nativeName
      type: object
      properties:
        isoCode:
          maxLength: 3
          minLength: 3
          type: string
        name:
          type: string
        nativeName:
          type: string
    CreateStateRequest:
      required:
        - code
        - name
        - nativeName
      type: object
      properties:
        code:
          type: string
        name:
          type: string
        nativeName:
          type: string
...
    State:
      required:
        - countryCode
        - code
        - name
        - nativeName
      type: object
      properties:
        countryCode:
          type: string
        code:
          type: string
        name:
          type: string
        nativeName:
          type: string
    StateWithCountry:
      required:
        - country
        - countryCode
        - code
        - name
        - nativeName
      type: object
      properties:
        country:
          $ref: "#/components/schemas/Country"
        countryCode:
          type: string
        code:
          type: string
        name:
          type: string
        nativeName:
          type: string
...
    UpdateCityRequest:
      required:
        - name
        - nativeName
      type: object
      properties:
        stateCode:
          type: string
          nullable: true
        name:
          type: string
        nativeName:
          type: string
...
    UpdateCountryRequest:
      required:
        - name
        - nativeName
      type: object
      properties:
        name:
          type: string
        nativeName:
          type: string
    UpdateStateRequest:
      required:
        - name
        - nativeName
      type: object
      properties:
        name:
          type: string
        nativeName:
          type: string
...
  securitySchemes:
    Bearer:
      type: http
      description: JWT Bearer authentication
      scheme: Bearer
    ConcertsCognitoAuthorizer:
      type: apiKey
      name: Authorization
      in: header
      x-amazon-apigateway-authtype: cognito_user_pools
      x-amazon-apigateway-authorizer:
        type: cognito_user_pools
        providerARNs:
          - Fn::ImportValue:
              Fn::Sub: '${BaseProductName}-cognito-authorizer-user-pool'
...